Prepare my files to include to main
Resolve #2
