+++ InfiniteLoop.h
@@ -1,0 +1,10 @@
+#pragma once
+#define _CRT_SECURE_NO_WARNINGS
+#define MAXSIZE 20
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <math.h>
+#include "InfiniteLoopFunctions.h"
+
+void InfiniteLoop(char Flies[][40], int AmountOfFlies, char InfiniteLoopSrtings[][100], int *AmountOfInfiniteMessages);
